COM-2727 - fix reviews
