+++ src/components/EventCell/index.tsx
@@ -10,16 +10,8 @@ import CameraAccessModal from '../modals/CameraAccessModal';
 import { COPPER, WHITE } from '../../styles/colors';
 import { hitSlop, ICON } from '../../styles/metrics';
 import styles, { eventCellStyleType } from './styles';
-import {
-  eventReducer,
-  initialState,
-  SET_EVENT_INFOS,
-  SET_TIMESTAMPED_INFOS,
-  cellReducer,
-  initialCellStyle,
-  SET_INTERNAL_HOUR_INFOS,
-  SET_INTERVENTION_INFOS,
-} from './reducers';
+import { eventReducer, initialState, SET_EVENT_INFOS, SET_TIMESTAMPED_INFOS } from './reducers/events';
+import { cellReducer, initialCellStyle, SET_INTERNAL_HOUR_INFOS, SET_INTERVENTION_INFOS } from './reducers/cells';
 
 interface TimeStampingProps {
   event: EventType,
@@ -51,8 +43,9 @@ const EventCell = ({ event }: TimeStampingProps) => {
   }, [event.histories]);
 
   useEffect(() => {
-    if (eventInfos.type === INTERVENTION) cellInfosDispatch({ type: SET_INTERVENTION_INFOS, payload: { eventInfos } });
-    else cellInfosDispatch({ type: SET_INTERNAL_HOUR_INFOS, payload: { eventInfos } });
+    const payload = eventInfos;
+    if (eventInfos.type === INTERVENTION) cellInfosDispatch({ type: SET_INTERVENTION_INFOS, payload });
+    else cellInfosDispatch({ type: SET_INTERNAL_HOUR_INFOS, payload });
   }, [eventInfos]);
 
   useEffect(() => setStyle(
@@ -135,7 +128,7 @@ const EventCell = ({ event }: TimeStampingProps) => {
             {!!eventInfos.endDate &&
               <Text style={style.eventInfo}> - {CompaniDate(eventInfos.endDate).format('HH:mm')}</Text>}
           </View>
-          <Text style={style.eventInfo}>{eventInfos.address.toLocaleLowerCase()}</Text>
+          <Text style={style.eventInfo}>{eventInfos.customer.address.toLocaleLowerCase()}</Text>
         </View>
         {!eventInfos.endDateTimeStamp && eventInfos.type === INTERVENTION &&
         <TouchableOpacity hitSlop={hitSlop} style={style.iconContainer}
